Move make_test() to meta.py
